[Decomp][Pet] Cohesion split: extract spells/talents to PetSpells.cpp
